Fix confusing/incorrect cache lookup wording in the TLB section.
The current wording seems to imply that caches are virtually indexed and
virtually tagged and do not require physical address, hence TLB lookup's
result.

However they are most commonly PIPT, and in any case always require the
physical address for the tag to avoid aliasing issues. The L1 set can
typically be selected from the virtual address only so the TLB lookup is
indeed - only partially - overlapped, but L2 and L3 need the physical
address prior to the lookup.